Made the display of the ActualRows column depending on wheter there are actuals to display or not.
Using NULL values to indicate the absensce of actuals internally (i.e. NOT using coalesce).
In the end (at the top of the code) check if there are any non-null values using COUNT. If not: the column is not displayed at all.
Note that the heading "ActualRows" will always be created, therefore the check of COUNT(ActualRows) OVER() needs to be > 1 to indicate the presence of REAL values (besides the heading).